backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow==11.3.0
python-decouple==3.8
psycopg2-binary==2.9.10
django-filter==23.3
drf-yasg==1.21.7
channels[daphne]==4.1.0
channels-redis==4.2.0
django-redis==5.4.0
redis==5.0.1
whitenoise==6.6.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0